Allow more general user modification of the light of the 2 stars.
This modification adds two extra array arguments, intended to support
much more flexible models of spots (e.g. to fit out of eclipse
modulations) by allowing the fractional change in the light of both
stars to be specified as a function of time.

The C API is broken by this change.  Python API is not, the subroutine
simply accepts optional extra keyword arguments.