When implementing Newton's method, complete the terminating iteration.
Should improve precision of the solutions (to be below the requested
error bound rather than flopping around it) at relatively minor cost.

I didn't do this originally just to save having the lines doing the
calculation appearing twice (inside and outside the loop) but there are
some simplifications that can be made at termination such as not
calculating quantities that are not used later.