Added beaming to the light curve calculation.
This shares the total radial velocity semiamplitude divided by the
speed of light with the light travel time calculation, the quantity
previously called CLTT in the parameter vector, which has now been
renamed to KTOTC to better reflect the shared role.  A flag NOLTT was
added to allow the light travel time correction to be disabled
independently of the beaming.

Two new parameters BEAM1 and BEAM2 were added to the parameter vector
to specify the beaming coefficients for the two stars, defined here as
alpha_beam = (3 - d log F_nu / d log nu) / 4 where the bolometric
value is unity and zero produces no beaming.