Added vectorized FMA intrinsics version of elliptic integrals K and E.
For x86-64 and double precision only, compiled in if the compiler
defines __amd64 and __FMA__ such as gcc with -mfma.  This was my
original intent for the dellke function but it wasn't worth doing
with just SSE2 instructions and I ended up waiting for FMA3 to
become widely available before I got round to doing it.

The performance benefit still isn't much, about 10% under test on my
workstation (Core i7 10700K).