Fixed issues caused by removal of PyArray_XDECREF_ERR in numpy 1.24.
The removal of this macro broke all of my modules on new versions
of numpy.  Fixed by writing a replacement enabled by an ifdef if
the original macro isn't defined, which does it the currently
recommended way (which is not backwards compatible so I couldn't
just replace all uses of the macro).